+++ test/schema.test.js
@@ -1,0 +1,10 @@
+var schemata = require('../')
+
+describe('#schema', function() {
+
+  it('schema should be empty for a default schemata', function() {
+    var empty = schemata()
+    empty.schema.should.eql({})
+  })
+
+})
